+++ src/ts-tooling.ts
@@ -50,10 +50,12 @@ import {StringFactory} from "./utils/string.factory";
 export {StringFactory} from "./utils/string.factory";
 import {ClassValidator} from './utils/class.validator';
 export {ClassValidator} from './utils/class.validator';
+import {ObjectFactory} from "./utils/object.factory";
+export {ObjectFactory} from "./utils/object.factory";
 
 export default {
     DateTime, TimeSpan, Guid, Byte, ByteStream,
     Dictionary, ListSortOrder,
     using, create, createWithFactory, EventHandler, ReactiveStore, BackgroundWorker,
-    LZCompression, StopWatch, NumberFactory, StringFactory, ClassValidator
+    LZCompression, StopWatch, NumberFactory, StringFactory, ClassValidator, ObjectFactory
 };