+++ src/type.extensions.ts
@@ -1,7 +1,6 @@
 import './primitive/number.extension';
 import './primitive/string.extension';
 import './primitive/list.extension';
-// import './primitive/object.extension';
 
 import {DateTime} from './complex/date.time';
 import {ListSortOrder} from './primitive/list.sort.order.enum';