+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 2.2.0
 
 * Reactive Store Wrap BehaviorSubject to prevent the User to change the Last Value and emit Values with next Function
+* add ObjectFactory with Circular Detection
 
 ### 2.1.3